+++ src/main/webapp/resources/js/components/alerts/ErrorAlert.tsx
@@ -4,16 +4,18 @@
 
 import React from "react";
 import { Alert } from "antd";
+import { AlertProps } from "antd/lib/alert";
 
 /**
  * Stateless UI component for displaying an [antd error Alert]{@link https://ant.design/components/alert/}
  *
- * @param {string} message - Text to display in alert
- * @param {string} description - Optional description
- * @param {object} props - any other props that are passed
- * @returns {Element} - Returns an antd error 'Alert' component
+ * @returns {JSX.Element} - Returns an antd error 'Alert' component
  */
-export function ErrorAlert({ message, description, ...props }) {
+export function ErrorAlert({
+  message,
+  description,
+  ...props
+}: AlertProps): JSX.Element {
   return (
     <Alert
       type="error"