+++ src/main/webapp/resources/js/types/irida/index.d.ts
@@ -1,4 +1,5 @@
 import { GenomeAssembly } from "../../apis/samples/samples";
+import { Restriction } from "../../utilities/restriction-utilities";
 
 export = IRIDA;
 export as namespace IRIDA;
@@ -146,7 +147,7 @@ declare namespace IRIDA {
     latitude: string;
     longitude: string;
     projects: Project[];
-    sequenceFiles: any[]; // TODO (Josh - 6/7/22): FLush this out
+    sequenceFiles: SequenceFile[];
     defaultSequencingObject: SequencingObject;
     defaultGenomeAssembly: GenomeAssembly;
     sampleName: string;
@@ -233,4 +234,12 @@ declare namespace IRIDA {
   }
 
   type UserMinimal = Pick<User, "name" | "id">;
+
+  export interface MetadataField {
+    id?: number;
+    fieldKey?: string;
+    label: string;
+    type?: string;
+    restriction: Restriction;
+  }
 }